first verion of unify devices clases
